Refactored common elements in the online emulator code into a base class
