+++ frontend/src/save-formats/OnlineEmulators/Emulators/Snes9x.js
@@ -2,46 +2,39 @@
 SNES9x save states appear to store the game's SRAM data after the magic "SRA:<size>:"
 */
 
-import SaveFilesUtil from '../../../util/SaveFiles';
+import EmulatorBase from './EmulatorBase';
 import Util from '../../../util/util';
 
 const MAGIC = 'SRA:'; // Magic begins with this string
 const SIZE_END_MAGIC = ':'; // Then we have the size of the SRAM data in bytes, which is terminated with this string
 const MAGIC_ENCODING = 'US-ASCII';
 
-function getRawArrayBufferFromSaveStateArrayBuffer(emulatorSaveStateArrayBuffer) {
-  try {
-    const magicOffset = Util.findMagic(emulatorSaveStateArrayBuffer, MAGIC, MAGIC_ENCODING);
-    const sizeBeginOffset = magicOffset + MAGIC.length;
-    const sizeEndOffset = Util.findMagic(emulatorSaveStateArrayBuffer, SIZE_END_MAGIC, MAGIC_ENCODING, sizeBeginOffset);
+export default class Snes9xSaveStateData extends EmulatorBase {
+  static getRawArrayBufferFromSaveStateArrayBuffer(emulatorSaveStateArrayBuffer) {
+    try {
+      const magicOffset = Util.findMagic(emulatorSaveStateArrayBuffer, MAGIC, MAGIC_ENCODING);
+      const sizeBeginOffset = magicOffset + MAGIC.length;
+      const sizeEndOffset = Util.findMagic(emulatorSaveStateArrayBuffer, SIZE_END_MAGIC, MAGIC_ENCODING, sizeBeginOffset);
 
-    const magicSizeDecoder = new TextDecoder(MAGIC_ENCODING);
-    const saveSizeString = magicSizeDecoder.decode(emulatorSaveStateArrayBuffer.slice(sizeBeginOffset, sizeEndOffset));
-    const saveSize = parseInt(saveSizeString, 10);
+      const magicSizeDecoder = new TextDecoder(MAGIC_ENCODING);
+      const saveSizeString = magicSizeDecoder.decode(emulatorSaveStateArrayBuffer.slice(sizeBeginOffset, sizeEndOffset));
+      const saveSize = parseInt(saveSizeString, 10);
 
-    const rawSaveBeginOffset = sizeEndOffset + SIZE_END_MAGIC.length;
-    const rawSaveEndOffset = rawSaveBeginOffset + saveSize;
+      const rawSaveBeginOffset = sizeEndOffset + SIZE_END_MAGIC.length;
+      const rawSaveEndOffset = rawSaveBeginOffset + saveSize;
 
-    return emulatorSaveStateArrayBuffer.slice(rawSaveBeginOffset, rawSaveEndOffset);
-  } catch (e) {
-    throw new Error('This does not appear to be a SNES9x save state file', e);
+      return emulatorSaveStateArrayBuffer.slice(rawSaveBeginOffset, rawSaveEndOffset);
+    } catch (e) {
+      throw new Error('This does not appear to be a SNES9x save state file', e);
+    }
   }
-}
 
-export default class Snes9xSaveStateData {
   static createFromSaveStateData(emulatorSaveStateArrayBuffer) {
-    const rawArrayBuffer = getRawArrayBufferFromSaveStateArrayBuffer(emulatorSaveStateArrayBuffer);
-
-    return new Snes9xSaveStateData(emulatorSaveStateArrayBuffer, rawArrayBuffer);
+    return super.createFromSaveStateData(emulatorSaveStateArrayBuffer, undefined, Snes9xSaveStateData);
   }
 
   static createWithNewSize(emulatorSaveStateData, newSize) {
-    // The user's emulator etc may require a different file size than the "true" size.
-    // We need to make sure that if the user resizes multiple times they don't lose data.
-    const originalRawArrayBuffer = getRawArrayBufferFromSaveStateArrayBuffer(emulatorSaveStateData.getEmulatorSaveStateArrayBuffer());
-    const newRawSaveData = SaveFilesUtil.resizeRawSave(originalRawArrayBuffer, newSize);
-
-    return new Snes9xSaveStateData(emulatorSaveStateData.getEmulatorSaveStateArrayBuffer(), newRawSaveData);
+    return super.createWithNewSize(emulatorSaveStateData, newSize, Snes9xSaveStateData);
   }
 
   static getRawFileExtension() {
@@ -54,18 +47,5 @@ export default class Snes9xSaveStateData {
 
   static fileSizeIsRequiredToConvert() {
     return false;
-  }
-
-  constructor(emulatorSaveStateArrayBuffer, rawArrayBuffer) {
-    this.emulatorSaveStateArrayBuffer = emulatorSaveStateArrayBuffer;
-    this.rawArrayBuffer = rawArrayBuffer;
-  }
-
-  getRawArrayBuffer() {
-    return this.rawArrayBuffer;
-  }
-
-  getEmulatorSaveStateArrayBuffer() {
-    return this.emulatorSaveStateArrayBuffer;
   }
 }
